Remove all coupling to Spotify from Scry package
